Add a directive to emit literal typescript
This solves the problem of a struct field being marshalled into a tuple.
It could also be acheived by putting frontmatter into tygo.yaml, but that
separates the TS definition from the type. Ideally we would like to say
  `tstype:"[a:number, b:number,c:string]"`
in the field tag, but commas are inescapably reserved for tag delimiters.